+++ src/diff.test.ts
@@ -1,11 +1,11 @@
 import { JSDOM } from 'jsdom'
 import { Options } from './config'
-import { CompareNodesResult, visualDomDiff } from './diff'
-import { areNodesEqual, isElement, isText } from './util'
+import { visualDomDiff } from './diff'
+import { areNodesEqual, charForNodeName, isElement, isText } from './util'
 
 jest.setTimeout(2000)
 
-const headings = new Set(['H1', 'H2', 'H3', 'H4', 'H5', 'H6'])
+const pChar = charForNodeName('P')
 
 function fragmentToHtml(documentFragment: DocumentFragment): string {
     return Array.from(documentFragment.childNodes).reduce(
@@ -202,25 +202,25 @@ test.each<[string, Node, Node, string, Options | undefined]>([
         undefined
     ],
     [
-        'a \\0 character replaces a paragraph',
+        'a character replaces a paragraph',
         (() => {
             const p = document.createElement('P')
             p.textContent = 'Test'
             return p
         })(),
-        document.createTextNode('\0Test'),
-        '<del class="vdd-removed"><p>Test</p></del><ins class="vdd-added">\0Test</ins>',
+        document.createTextNode(`${pChar}Test`),
+        `<del class="vdd-removed"><p>Test</p></del><ins class="vdd-added">${pChar}Test</ins>`,
         undefined
     ],
     [
-        'paragraph replaces a \\0 character',
-        document.createTextNode('\0Test'),
+        'paragraph replaces a character',
+        document.createTextNode(`${pChar}Test`),
         (() => {
             const p = document.createElement('P')
             p.textContent = 'Test'
             return p
         })(),
-        '<del class="vdd-removed">\0Test</del><ins class="vdd-added"><p>Test</p></ins>',
+        `<del class="vdd-removed">${pChar}Test</del><ins class="vdd-added"><p>Test</p></ins>`,
         undefined
     ],
     [
@@ -423,54 +423,15 @@ test.each<[string, Node, Node, string, Options | undefined]>([
         undefined
     ],
     [
-        'custom compareNodes option - ignore attributes',
-        htmlToFragment('<div><br><img src="image.jpg"></div>'),
-        htmlToFragment('<div><hr><img src="image.png"></div>'),
-        '<div><del class="vdd-removed"><br></del><ins class="vdd-added"><hr></ins>' +
-            '<ins class="vdd-modified"><img src="image.png"></ins></div>',
-        {
-            compareNodes(node1: Node, node2: Node) {
-                return areNodesEqual(node1, node2)
-                    ? CompareNodesResult.IDENTICAL
-                    : node1.nodeName === node2.nodeName
-                    ? CompareNodesResult.SIMILAR
-                    : CompareNodesResult.DIFFERENT
-            }
-        }
-    ],
-    [
-        'custom compareNodes option - UL similar to OL',
-        htmlToFragment('<div><ol><li>Item</li><li>One</li></ol></div>'),
-        htmlToFragment('<div><ul><li>Item</li><li>Two</li></ul></div>'),
-        '<div><ins class="vdd-modified"><ul><li>Item</li>' +
-            '<li><del class="vdd-removed">One</del><ins class="vdd-added">Two</ins></li></ul></ins></div>',
-        {
-            compareNodes(node1: Node, node2: Node) {
-                return (node1.nodeName === 'UL' && node2.nodeName === 'OL') ||
-                    (node1.nodeName === 'OL' && node2.nodeName === 'UL')
-                    ? CompareNodesResult.SIMILAR
-                    : undefined
-            }
-        }
-    ],
-    [
-        'custom compareNodes option - all headings are identical',
+        'a table with a changed cell and text added after',
         htmlToFragment(
-            '<h1>Heading 1</h1><h2>Heading 2</h2><h3>Heading 3</h3><h1>Different</h1>'
+            `<p>This homepage is a bit unusual but I wouldn't worry about it. Nor would I tell the authorities, EVER!</p><table border="1" style="border-collapse:collapse;width:360px;"><tbody><tr><td style="width:253px;">scsac</td><td style="width:253px;">ascasc</td><td style="width:253px;">ascsac</td></tr><tr><td style="width:253px;">ascasc</td><td style="width:253px;">ascasc</td><td style="width:253px;">ascasc</td></tr><tr><td style="width:253px;"><br></td><td style="width:253px;">ascasc</td><td style="width:253px;"><br></td></tr></tbody></table><p><br></p>`
         ),
         htmlToFragment(
-            '<h4>Heading 1</h4><h5>Heading 2</h5><h6>Heading 3</h6><p>Different</p>'
+            `<p>This homepage is a bit unusual but I wouldn't worry about it. Nor would I tell the authorities, EVER!</p><table border="1" style="border-collapse:collapse;width:360px;"><tbody><tr><td style="width:253px;">scsac</td><td style="width:253px;">ascasc</td><td style="width:253px;">ascsac</td></tr><tr><td style="width:253px;">ascasc</td><td style="width:253px;">ascasc</td><td style="width:253px;">ascasc</td></tr><tr><td style="width:253px;">wdqdqwd</td><td style="width:253px;">ascasc</td><td style="width:253px;"><br></td></tr></tbody></table><p>sacascascas</p><p>sacasc</p><p>ascasc</p><p><br></p>`
         ),
-        '<h4>Heading 1</h4><h5>Heading 2</h5><h6>Heading 3</h6>' +
-            '<del class="vdd-removed"><h1>Different</h1></del><ins class="vdd-added"><p>Different</p></ins>',
-        {
-            compareNodes(node1: Node, node2: Node) {
-                return headings.has(node1.nodeName) &&
-                    headings.has(node2.nodeName)
-                    ? CompareNodesResult.IDENTICAL
-                    : undefined
-            }
-        }
+        `<p>This homepage is a bit unusual but I wouldn't worry about it. Nor would I tell the authorities, EVER!</p><table border="1" style="border-collapse:collapse;width:360px;"><tbody><tr><td style="width:253px;">scsac</td><td style="width:253px;">ascasc</td><td style="width:253px;">ascsac</td></tr><tr><td style="width:253px;">ascasc</td><td style="width:253px;">ascasc</td><td style="width:253px;">ascasc</td></tr><tr><td style="width:253px;"><del class="vdd-removed"><br></del><ins class="vdd-added">wdqdqwd</ins></td><td style="width:253px;">ascasc</td><td style="width:253px;"><br></td></tr></tbody></table><ins class="vdd-added"><p>sacascascas</p><p>sacasc</p><p>ascasc</p></ins><p><br></p>`,
+        undefined
     ]
 ])(
     '%s',